+++ app/controllers/authentications_controller.rb
@@ -26,7 +26,7 @@ class AuthenticationsController < ApplicationController
 
   # This wrapper of the oauth route exists to do reauth before adding
   def add
-    redirect_to "/auth/#{params[:provider]}"
+    redirect_to "/auth/#{params[:provider]}?add=true"
   end
 
 end
